Ready to deploy

+++ API/DTOs/ClientDTOs/StationClientDTO.cs
@@ -9,8 +9,8 @@ namespace API.DTOs.ClientDTOs
     {
         public int Id { get; set; }
         public string StationName { get; set; }
-        public decimal GegrLat { get; set; }
-        public decimal GegrLon { get; set; }
+        public double GegrLat { get; set; }
+        public double GegrLon { get; set; }
         public string AddressStreet { get; set; }
         public ICollection<SensorDTO> Sensors { get; set; }
     }
