+++ API/DTOs/StationDTO.cs
@@ -9,8 +9,8 @@ namespace API.DTOs
     {
         public int Id { get; set; }
         public string StationName { get; set; }
-        public decimal GegrLat { get; set; }
-        public decimal GegrLon { get; set; }
+        public double GegrLat { get; set; }
+        public double GegrLon { get; set; }
         public CityDTO City { get; set; }
         public string AddressStreet { get; set; }
     }
