+++ API/Entities/Station.cs
@@ -9,8 +9,8 @@ namespace API.Entities
     {
         public int Id { get; set; }
         public string StationName { get; set; }
-        public decimal GegrLat { get; set; }
-        public decimal GegrLon { get; set; }
+        public double GegrLat { get; set; }
+        public double GegrLon { get; set; }
         public int CityId { get; set; }
         public virtual City City { get; set; }
         public string AddressStreet { get; set; }
